+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   into the APPLgrid format
 - added support for [EKO](https://eko.readthedocs.io/en/latest/)'s new file
   formats (old file format is still supported) in `pineappl evolve`
+- `pineappl import` now also imports fastNLO's metadata:
+  `fastNLOTable::GetDimLabels()` is stored as the value of the key
+  `x{1,2,...}_labels` and `fastNLOTable::GetScDescr()` is stored as the value of
+  the key `fastnlo_scenario`
 
 ### Changed